pyproject.toml
[project]
name = "ai-resume-bot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "chromadb>=1.3.2",
    "langchain>=0.3.0",
    "langchain-openai>=0.2.0",
    "langchain-community>=0.3.0",
    "langchain-text-splitters>=0.3.0",
    "python-dotenv>=1.2.1",
    "gradio>=5.0.0",
    "openai>=1.0.0",
    "pypdf>=5.0.0",
    "huggingface-hub>=1.0.1",
]
